add internet check to openfe
